rbac: manifests: prune roles
per https://kubernetes.io/docs/reference/access-authn-authz/rbac/:

"""
In the RBAC API, a role contains rules that represent a set of permissions.
Permissions are purely additive (there are no “deny” rules).
A role can be defined within a namespace with a Role, or cluster-wide with
a ClusterRole.
[...]
A ClusterRole can be used to grant the same permissions as a Role,
but because they are cluster-scoped, they can also be used to grant
access to:

- cluster-scoped resources (like nodes)
- non-resource endpoints (like “/healthz”)
- namespaced resources (like pods) across all namespaces (needed
  to run kubectl get pods --all-namespaces, for example)
"""

So we actually need only ClusterRoles, and their binding.

Signed-off-by: Francesco Romani <fromani@redhat.com>