manifest: drop unneeded 'bind' verb
Testing of the CSV revealed that the 'bind' verb doesn't
pass the validation on OCP - it is not recognized among the
supported verbs.

Search on the historical records doesn't reveal why this verb
was added, and it is not found after a (quick) search in the
k8s/okd docs.

For these reasons, we just drop it.

Signed-off-by: Francesco Romani <fromani@redhat.com>