functests: bootstrap and initial batch of tests
We start testing the deployment of the common templates.
This patch also paves the road for future tests, e.g. about the
validator and the labeller.
We include an initial trivial test for the validator as reference.

The functests assume a running OCP/OKD cluster, see README.md

Running time:

* [01-test-operator-running.sh                                     ] OK
* [02-test-deploy-templates-openshift.sh                           ] OK
* [03-test-deploy-templates-different-namespace.sh                 ] OK
* [04-test-deploy-templates-operator-namespace.sh                  ] OK
* [05-test-deploy-template-validator.sh                            ] OK
real 167.37
user 14.94
sys 2.67

Signed-off-by: Francesco Romani <fromani@redhat.com>