ssp: labeller: expects scheduled pods
Previously, the condition reporting logic could have reported
Avaliable=True
If 0 == Scheduled == Available, which is wrong. To be available
we need all scheduled pods online, but also _some_ scheduled pods.

We don't have reports from field to learn if this issue actually
happened or if it is only possible in theory, but still is a bug,
so this patch fixes it.

Signed-off-by: Francesco Romani <fromani@redhat.com>