operator: implement circuitbreaker
Implement circuitbreaker-like behaviour for the SSP operator.
With is change is now possible to *temporarily* make the operator ignore
one or more operands.
This is intended to be used as extrema ratio parachute
to troubleshoot issue in clusters, when everything else failed and the
state has to be recovered, or at very least investigated, somehow.

see `TROUBLESHOOTING.md` to learn about how to use this
functionality.

Signed-off-by: Francesco Romani <fromani@redhat.com>