operator: implement circuitbreaker
Implement circuitbreaker-like behaviour for the SSP operator.
With is change is now possible to *temporarily* make the operator ignore
one or more operands.
This is intended to be used as extrema ratio parachute
to troubleshoot issue in clusters, when everything else failed and the
state has to be recovered, or at very least investigated, somehow.

The initial idea was to add an annotation to each operand CR.
Unfortunately, the ansible-operator doesn't propagate all the CR
metadata but just `name` and `namespace`, see
https://github.com/operator-framework/operator-sdk/blob/v0.10.x/pkg/ansible/runner/runner.go#L236

It could maybe be possible to fetch each operand CR from its playbook,
but this made harder to create a reusable role, requiring changes to all
the roles, something we'd like to avoid to make maintenance easy.

Instead, we move the annotation to the operator pod.
This is actually helpful to enforce the ephemeral nature of the circuitbreaker,
and the fact it must be default off, opt-in.

Signed-off-by: Francesco Romani <fromani@redhat.com>