Custom image name prefix for the CSV generator
Let the user specify a custom image name prefix
running the CSV generator.
The default value is an empty string.

This enables downstream consumers to
build ssp related images setting a
custom common prefix on all the images
names.
A better option would be to pass the full
pull URL for each image instead of just
the specific tag and a common prefix
but it would be a wider change that
needs more reobust testing.

Fixes: https://bugzilla.redhat.com/1757798

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>