node-labeller: use snake-case internally, again.
When the ansible operator creates the playbook parameters,
it takes all the attributes under `spec:` and creates
variables out of them.
However, the variable names are not taken verbatim, but
they are translated to snake_case:
https://github.com/operator-framework/operator-sdk/blob/v0.11.x/pkg/ansible/runner/runner.go#L259

So the role variables should always use the `snake_case` variant
automatically created, not the `camelCase` exposed version.
Out of necessity, this commit undoes most of the changes
done in 83528c8 because we had the following bug:

1. HCO wants to set `useKVM=false` when running on its CI
2. the ansible operator translates the variable to `use_kvm=false`
   and passes the parameter to the playbook
3. the playbook runs, and sets the defaults, including `useKVM=true`
4. the playbook consumes the `useKVM` (camelCase) variable, whose
   defaults has been set and never overriden
5. the `use_kvm` (snake_case) variable, which holds the correct value
   the user set (in the CR spec.useKVM!) is ignored entirely.

So the only fix is to expose fields to kubernetes in camelCase (`useKVM`)
but to consume their snake_case counterparts (`use_kvm`) in the roles.

Signed-off-by: Francesco Romani <fromani@redhat.com>