decreased timeout time in "wait" tasks
This change should prevent uncontrollable "freeze" of ssp operator.
This "freeze" happens in this scenario(mostly during tests): e.g. template-validator
is deployed. Then ansible moves to "Wait for the template-validator to start" task.
But when user (or script) deletes template-validator CR, operator keeps trying to test
if validator is online or not (in this case for more than 3000 seconds) and then operator fails
and whole container is restarted. This patch decrease this time to only 60 seconds and prevents restarts
of container by adding flag ignore_errors to "wait" tasks (for now only validator + node-labeller have this task)

Signed-off-by: Karel Šimon <ksimon@redhat.com>