Make deployment template compatible with 'sha256:' tags
Image pull URLs can contain ':tag' or '@sha256:digest' so we
have to make deployment templates compatible with both the syntax
instead of hardcoding ':' as the separator and always assuming
that the env variable always contains a plain tag.
Now the templates will check if the tag variable starts with
'sha256:' and in that case it will use '@' as the delimiter
as other operators are doing.