fix labeller when no kvm device is available on the node
When node doesn't have access to kvm, labeller stays in pending state. This commit
removes the requirement for devices.kubevirt.io/kvm: "1". It adds new check which control
if pod has access to kvm device. If yes, everything goes as usual (binaries are copied, node is updated with new labels).
If kvm is not available, containers exit with 0 and no labells are added.

Signed-off-by: Karel Simon <ksimon@redhat.com>