Addresses: https://bugzilla.redhat.com/show_bug.cgi?id=1853133
The k8s module's 'apply' options requires a version of the python2-openshift package that is not available downstream.
This commit replaces the 'apply' option with a merge/json patch (both are tried) that lifts this requirement.

This commit also makes sure the operator uses the 'templates' kind to prevent collisions with other resources, see: https://github.com/operator-framework/operator-sdk/issues/1380

Signed-off-by: Omer Yahud <oyahud@redhat.com>