Fixed a flakey explainable CRD test.
I couldn't find out why the CRD is not available for 'oc explain', I have tried
waiting for conditions on the CRD that mark it as ready but that did not help.

I eventually resorted to adding an arbitrary sleep before executing the test and that
seemed to solve the issue.

Signed-off-by: Omer Yahud <oyahud@redhat.com>